move recommended app permissions to the system-data package
